Fix pre-existing build blockers and commit a fresh web build
flutter build web was failing before any of this branch's feature work,
independent of it:
- management_executive_dashboard.dart passed a non-existent onReorderItem
  parameter to ReorderableListView.builder instead of onReorder.
- report_catalog_service.dart / report_service.dart used const pw.TextStyle(...)
  in ~30 places; the locked pdf 3.12.0 package can't const-evaluate its
  TextStyle constructor on this Dart SDK, so every const usage failed
  compilation. Dropping const (verified with an isolated repro against the
  same package version) fixes it with no behavior change.
- ai_lead_score.dart: two num.clamp() results were assigned into double
  fields without a .toDouble() call.

Vercel's build script only ships whatever is already committed under
build/web/ (it does not run flutter build web), so source changes never
reached the live site regardless of branch. This commit is a freshly
built build/web from the fixed source so the deployed app actually
reflects the feature branch.

+++ lib/analytics/ai_lead_score.dart
@@ -244,7 +244,8 @@ abstract final class AiLeadScore {
     List<LandLeadLegalDocument> legalDocs,
   ) {
     final present = _presentDocTypes(legalDocs);
-    final raw = (present.length / _requiredDocTypes.length * 100).clamp(0, 100);
+    final raw =
+        (present.length / _requiredDocTypes.length * 100).clamp(0, 100).toDouble();
     final missing =
         _requiredDocTypes.where((t) => !present.contains(t)).toList();
     return AiScoreCategory(
@@ -372,7 +373,7 @@ abstract final class AiLeadScore {
 
     if (notes.isEmpty) notes.add('No risk indicators detected');
 
-    final raw = (100 - deductions).clamp(0, 100);
+    final raw = (100 - deductions).clamp(0, 100).toDouble();
     return AiScoreCategory(
       label: 'Risk Factors',
       weightPercent: 10,
